render_app/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1

# OpenAI + LangChain
langchain==0.2.12
langchain-core==0.2.35
langchain-pinecone==0.1.0
langchain-openai==0.1.3

# Pinecone (latest stable)
pinecone-client==3.2.2

tiktoken==0.7.0
requests==2.32.3
